Disable Elasticsearch's zen multicast discovery.
When multicast discovery is on, an instance will send UDP ping packets and
discover other instances within the cluster on the same subnet. Since we don't
name our clusters to differentiate them among containers, we want to turn this
off to avoid Elasticsearch instances on the same subnet creating a cluster
together.